# OpenAPI/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/routes:
    get:
      tags:
      - Routes
      summary: Get all routes
      description: "Retrieve routes with optional filtering by type, visibility, and\
        \ source"
      operationId: getAllRoutes
      parameters:
      - name: type
        in: query
        description: "Type of routes: noActivities, withActivities"
        required: false
        schema:
          type: string
      - name: from
        in: query
        description: "Filter by source (user, mural, etc)"
        required: false
        schema:
          type: string
      - name: id
        in: query
        description: Filter by source ID
        required: false
        schema:
          type: integer
          format: int64
      - name: visibility
        in: query
        description: "Filter by visibility: PUBLIC, PRIVATE, MURAL_PUBLIC, MURAL_SPECIFIC"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Routes retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteDtoInterface"
        "400":
          description: Invalid visibility type
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteDtoInterface"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteDtoInterface"
    post:
      tags:
      - Routes
      summary: Create new route
      description: Create a new route from an existing activity. Pass the activity
        ID in the route object's id field; it will be cleared and the route will be
        persisted with a new ID
      operationId: createRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Route"
        required: true
      responses:
        "200":
          description: Route created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RouteWithActivityDTO"
        "404":
          description: Activity not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RouteWithActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RouteWithActivityDTO"
  /api/routes/{routeId}/activities/:
    post:
      tags:
      - Routes
      summary: Add activities to route
      description: Add multiple activities to a route
      operationId: addActivitiesToRoute
      parameters:
      - name: routeId
        in: path
        description: Route ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
        required: true
      responses:
        "404":
          description: Route not found
          content:
            '*/*':
              schema:
                type: boolean
        "204":
          description: Activities added successfully
          content:
            '*/*':
              schema:
                type: boolean
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: boolean
...
  /api/routes/{id}/visibility:
    patch:
      tags:
      - Routes
      summary: Change route visibility
      description: "Update the visibility settings of a route. Request body should\
        \ contain 'visibility' (PUBLIC, PRIVATE, MURAL_PUBLIC, MURAL_SPECIFIC) and\
        \ optionally 'allowedMuralsList' as a comma-separated string in brackets for\
        \ MURAL_SPECIFIC visibility"
      operationId: changeVisibility
      parameters:
      - name: id
        in: path
        description: Route ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "404":
          description: Route not found
          content:
            '*/*':
              schema:
                type: boolean
        "204":
          description: Visibility changed successfully
          content:
            '*/*':
              schema:
                type: boolean
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: boolean
  /api/routes/visibility/mural:
    patch:
      tags:
      - Routes
      summary: Hide routes from mural
      description: Make routes not visible to a specific mural
      operationId: makeRoutesNotVisibleToMural
      parameters:
      - name: id
        in: query
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
        required: true
      responses:
        "204":
          description: Routes visibility updated
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
...
  /api/activities/{id}/visibility:
    patch:
      tags:
      - Activities
      summary: Change activity visibility
      description: "Update the visibility settings of an activity. Request body should\
        \ contain 'visibility' (PUBLIC, PRIVATE, MURAL_PUBLIC, MURAL_SPECIFIC) and\
        \ optionally 'allowedMuralsList' as a comma-separated string in brackets for\
        \ MURAL_SPECIFIC visibility"
      operationId: changeVisibility_1
      parameters:
      - name: id
        in: path
        description: Activity ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "404":
          description: Activity not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "200":
          description: Visibility changed successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
...
  /api/routes/{id}:
    get:
      tags:
      - Routes
      summary: Get route by ID
      description: Retrieve a route's information by its ID
      operationId: getRoute
      parameters:
      - name: id
        in: path
        description: Route ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Route not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RouteWithoutActivityDTO"
        "200":
          description: Route found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RouteWithoutActivityDTO"
    delete:
      tags:
      - Routes
      summary: Delete route
      description: Delete a route by its ID
      operationId: deleteRoute
      parameters:
      - name: id
        in: path
        description: Route ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Route not found
          content:
            '*/*':
              schema:
                type: boolean
        "204":
          description: Route deleted successfully
          content:
            '*/*':
              schema:
                type: boolean
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: boolean
  /api/routes/{id}/activities:
    get:
      tags:
      - Routes
      summary: Get activities in route
      description: Retrieve all activities assigned to a specific route
      operationId: getActivitiesAssignedToRoute
      parameters:
      - name: id
        in: path
        description: Route ID
        required: true
        schema:
          type: integer
          format: int64
      - name: mural
        in: query
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Route not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityOfRouteDTO"
        "200":
          description: Activities retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityOfRouteDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityOfRouteDTO"
  /api/routes/OwnedInMural:
    get:
      tags:
      - Routes
      summary: Get owned routes in mural
      description: Retrieve routes owned by the current user in a specific mural
      operationId: getOwnedRoutesInMural
      parameters:
      - name: muralId
        in: query
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Routes retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteWithoutActivityDTO"
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteWithoutActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteWithoutActivityDTO"
...
  /api/routes/{routeId}/activities/{activityId}:
    delete:
      tags:
      - Routes
      summary: Remove activity from route
      description: Remove an activity from a route
      operationId: removeActivityFromRoute
      parameters:
      - name: routeId
        in: path
        description: Route ID
        required: true
        schema:
          type: integer
          format: int64
      - name: activityId
        in: path
        description: Activity ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Route or activity not found
          content:
            '*/*':
              schema:
                type: boolean
        "204":
          description: Activity removed successfully
          content:
            '*/*':
              schema:
                type: boolean
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: boolean
...
components:
  schemas:
...
    Mural:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        code:
          type: string
        visibility:
          type: string
          enum:
          - PRIVATE
          - MURAL_SPECIFIC
          - MURAL_PUBLIC
          - PUBLIC
        owner:
          $ref: "#/components/schemas/User"
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        bannedUsers:
          type: array
          items:
            $ref: "#/components/schemas/User"
        thumbnail:
          type: array
          items:
            type: string
            format: byte
        banner:
          type: array
          items:
            type: string
            format: byte
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        name:
          type: string
        email:
          type: string
        memberMurals:
          type: array
          items:
            $ref: "#/components/schemas/Mural"
        roles:
          type: array
          items:
            type: string
        admin:
          type: boolean
    Coordinates:
      type: object
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
    Route:
      type: object
      properties:
        id:
          type: integer
          format: int64
        totalDistance:
          type: number
          format: double
        elevationGain:
          type: number
          format: double
        coordinates:
          type: array
          items:
            $ref: "#/components/schemas/Coordinates"
        name:
          type: string
        description:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        createdBy:
          $ref: "#/components/schemas/User"
      description: The route object must hold the target activity ID in its id field.
    Visibility:
      type: object
      properties:
        type:
          type: string
          enum:
          - PRIVATE
          - MURAL_SPECIFIC
          - MURAL_PUBLIC
          - PUBLIC
        allowedMurals:
          type: array
          items:
            type: integer
            format: int64
        public:
          type: boolean
        private:
          type: boolean
        allowedMuralsNonNull:
          type: array
          items:
            type: integer
            format: int64
        muralSpecific:
          type: boolean
        muralPublic:
          type: boolean
    ActivityOfRouteDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        user:
          type: integer
          format: int64
        summary:
          $ref: "#/components/schemas/ActivitySummary"
    ActivitySummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startTime:
          type: string
          format: date-time
        totalDistance:
          type: number
          format: double
        totalTime:
          type: integer
          format: int64
        elevationGain:
          type: number
          format: double
        averages:
          type: object
          additionalProperties:
            type: number
            format: double
        records:
          type: object
          additionalProperties:
            type: string
        upRecords:
          type: object
          additionalProperties:
            type: string
          writeOnly: true
        upAverages:
          type: object
          additionalProperties:
            type: number
            format: double
          writeOnly: true
    RouteWithActivityDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        totalDistance:
          type: number
          format: double
        elevationGain:
          type: number
          format: double
        visibility:
          $ref: "#/components/schemas/Visibility"
        coordinates:
          type: array
          items:
            $ref: "#/components/schemas/Coordinates"
        name:
          type: string
        description:
          type: string
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityOfRouteDTO"
        createdBy:
          type: integer
          format: int64
...
    NamedId:
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
          format: int64
...
    ActivityDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
        startTime:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/NamedId"
        route:
          $ref: "#/components/schemas/NamedId"
        dataPoints:
          type: array
          items:
            $ref: "#/components/schemas/DataPoint"
        streams:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        summary:
          $ref: "#/components/schemas/ActivitySummary"
        visibility:
          type: string
          enum:
          - PRIVATE
          - MURAL_SPECIFIC
          - MURAL_PUBLIC
          - PUBLIC
        allowedMurals:
          type: array
          items:
            type: integer
            format: int64
    DataPoint:
      type: object
      properties:
        time:
          type: string
          format: date-time
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
        ele:
          type: number
          format: double
        hr:
          type: integer
          format: int32
        cad:
          type: integer
          format: int32
        other:
          type: object
          additionalProperties:
            type: string
...
    RouteDtoInterface:
      type: object
    RouteWithoutActivityDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        totalDistance:
          type: number
          format: double
        elevationGain:
          type: number
          format: double
        visibility:
          $ref: "#/components/schemas/Visibility"
        coordinates:
          type: array
          items:
            $ref: "#/components/schemas/Coordinates"
        name:
          type: string
        description:
          type: string
        createdBy:
          type: integer
          format: int64